refactor(acceptance): land in-flight acceptance, artifact and review-stage work
Working-tree state committed by area.
