--- packages/opencorvus/src/architect/index.ts
@@ -1,2 +1,0 @@
-export { ArchitectAgent } from "./agent"
-export type { ArchitectArtifact, ArchitectContract, ArchitectDecisionKey } from "./types"
