--- packages/opencorvus/src/goal-workload-analyst/index.ts
@@ -1,3 +1,0 @@
-export { GoalWorkloadAnalystAgent } from "./agent"
-export type { GoalWorkloadArtifact, WorkloadBrief } from "./types"
-export { createGoalWorkloadOutputTools, type GoalWorkloadCollector } from "./output-tools"
