--- packages/opencorvus/src/intent-analysis/index.ts
@@ -1,9 +1,0 @@
-export { IntentAnalysisAgent } from "./agent"
-export type {
-  IntentAnalysisResult,
-  IntentClarification,
-  IntentClass,
-  IntentComplexity,
-  IntentSlot,
-  ClarificationPriority,
-} from "./types"
