+++ packages/opencorvus/src/work-artifact/profile-registry.ts
@@ -11,6 +11,18 @@ export const WORK_ARTIFACT_TOOL_IDS = [
   "work_artifact_deliver",
 ] as const
 
+/** Named so the delivery authority can match on it instead of a SQL string literal. */
+export const WORK_ARTIFACT_VALIDATE_TOOL_ID = "work_artifact_validate" as const
+/**
+ * Named for the same reason, and because prose that grants or withholds a Tool
+ * must not reach for it by array position: `delegate-agent.ts` addressed these
+ * as `WORK_ARTIFACT_TOOL_IDS[0]` / `[2]` / `[3]`, so reordering the array above
+ * would have rewritten what a sub-agent is told it may do, with no compile
+ * error and no test covering the sentence.
+ */
+export const WORK_ARTIFACT_INSPECT_TOOL_ID = "work_artifact_inspect" as const
+export const WORK_ARTIFACT_DELIVER_TOOL_ID = "work_artifact_deliver" as const
+
 export const WORK_ARTIFACT_PARENT_ONLY_TOOL_IDS = ["work_artifact_deliver"] as const
 export const WORK_ARTIFACT_PRESENTATION_TOOL_SCHEMA_REVISION = "work-artifact-presentation@1" as const
 export const OFFICE_PRESENTATION_PROFILE_LIMITS = Object.freeze({
@@ -74,16 +86,15 @@ const profiles = new Map<WorkArtifactProfileID, WorkArtifactProfile>([
   ],
 ])
 
+/**
+ * Only `require` had a caller. `all()` and `supports()` were registry-shaped
+ * API for a map with one entry: nothing outside this file ever asked for the
+ * whole set or for operation support.
+ */
 export const WorkArtifactProfileRegistry = Object.freeze({
-  all(): readonly WorkArtifactProfile[] {
-    return WORK_ARTIFACT_PROFILE_IDS.map((id) => profiles.get(id)!)
-  },
   require(id: WorkArtifactProfileID): WorkArtifactProfile {
     const profile = profiles.get(id)
     if (!profile) throw new Error(`Unknown qualified Work Artifact profile: ${id}`)
     return profile
-  },
-  supports(id: WorkArtifactProfileID, operation: WorkArtifactOperationID): boolean {
-    return profiles.get(id)?.operations.includes(operation) ?? false
   },
 })
